+++ tsconfig.cloudflare.json
@@ -7,6 +7,7 @@
     "app/**/.client/**/*",
     "workers/**/*",
     "worker-configuration.d.ts",
+    "lib/**/*",
   ],
   "compilerOptions": {
     "composite": true,
